web-crawling/requirements.txt
requests
beautifulsoup4
lxml
goose3
newspaper3k
trafilatura